+++ noodles-sam/src/record.rs
@@ -548,7 +548,7 @@ impl AlignmentRecord for Record {
     /// # Ok::<_, io::Error>(())
     /// ```
     fn alignment_span(&self) -> io::Result<u32> {
-        Ok(self.cigar().reference_len())
+        Ok(self.cigar().reference_len() as u32)
     }
 
     fn mapping_quality(&self) -> Option<MappingQuality> {